+++ Visual Studio Project/Test Console App/Program.cs
@@ -9,12 +9,6 @@ namespace Test_Console_App
         static Driver D;
         static void Main(string[] args)
         {
-
-        }
-
-        private static void D_DriverReady()
-        {
-           
         }
     }
 }