+++ Visual Studio Project/ZWaveJS.NET/Controller.cs
@@ -10,6 +10,13 @@ namespace ZWaveJS.NET
 
         }
 
+        public delegate void StatisticsUpdatedEvent(ControllerStatistics Statistics);
+        public event StatisticsUpdatedEvent StatisticsUpdated;
+        internal void Trigger_StatisticsUpdated(ControllerStatistics Statistics)
+        {
+            this.statistics = Statistics;
+            StatisticsUpdated?.Invoke(Statistics);
+        }
 
         public delegate void InclusionAbortedEvent();
         public event InclusionAbortedEvent InclusionAborted;
@@ -349,6 +356,8 @@ namespace ZWaveJS.NET
         public bool supportsTimers { get; internal set; }
         [Newtonsoft.Json.JsonProperty]
         public bool isHealNetworkActive { get; internal set; }
+        [Newtonsoft.Json.JsonProperty]
+        public ControllerStatistics statistics { get; internal set; }
 
     }
 }