+++ Visual Studio Project/ZWaveJS.NET/Driver.cs
@@ -12,7 +12,8 @@ namespace ZWaveJS.NET
     {
         internal static WebsocketClient Client;
         internal static Dictionary<Guid, Action<JObject>> Callbacks;
-        private Dictionary<string, Action<JObject>> EventMap;
+        private Dictionary<string, Action<JObject>> NodeEventMap;
+        private Dictionary<string, Action<JObject>> ControllerEventMap;
         private static int SchemaVersionID = 14;
         internal static CustomBooleanJsonConverter BoolConverter;
 
@@ -43,7 +44,41 @@ namespace ZWaveJS.NET
 
         private void MapNodeEvents()
         {
-            EventMap.Add("value updated", (JO) =>
+            NodeEventMap.Add("statistics updated", (JO) =>
+            {
+                int NID = JO.SelectToken("event.nodeId").Value<int>();
+                NodeStatistics NS = JsonConvert.DeserializeObject<NodeStatistics>(JO.SelectToken("event.statistics").ToString());
+                ZWaveNode N = this.Controller.Nodes.Get(NID);
+                N.Trigger_StatisticsUpdated(NS);
+            });
+
+            NodeEventMap.Add("firmware update finished", (JO) =>
+            {
+                int NID = JO.SelectToken("event.nodeId").Value<int>();
+                int Status = JO.SelectToken("event.status").Value<int>();
+                int Wait = 0;
+
+                if(JO.SelectToken("event.waitTime") != null)
+                {
+                    Wait = JO.SelectToken("event.waitTime").Value<int>();
+                }
+
+                ZWaveNode N = this.Controller.Nodes.Get(NID);
+                N.Trigger_FirmwareUpdateFinished(Status, Wait);
+                
+            });
+
+
+            NodeEventMap.Add("firmware update progress", (JO) =>
+            {
+                int NID = JO.SelectToken("event.nodeId").Value<int>();
+                int Sent = JO.SelectToken("event.sentFragments").Value<int>();
+                int Total = JO.SelectToken("event.totalFragments").Value<int>();
+                ZWaveNode N = this.Controller.Nodes.Get(NID);
+                N.Trigger_FirmwareUpdateProgress(Sent, Total);
+            });
+
+            NodeEventMap.Add("value updated", (JO) =>
             {
                 int NID = JO.SelectToken("event.nodeId").Value<int>();
                 JObject IJO = JO.SelectToken("event.args").Value<JObject>();
@@ -51,7 +86,7 @@ namespace ZWaveJS.NET
                 N.Trigger_ValueUpdated(IJO);
             });
 
-            EventMap.Add("value added", (JO) =>
+            NodeEventMap.Add("value added", (JO) =>
             {
                 int NID = JO.SelectToken("event.nodeId").Value<int>();
                 JObject IJO = JO.SelectToken("event.args").Value<JObject>();
@@ -59,7 +94,7 @@ namespace ZWaveJS.NET
                 N.Trigger_ValueUpdated(IJO);
             });
 
-            EventMap.Add("value notification", (JO) =>
+            NodeEventMap.Add("value notification", (JO) =>
             {
                 int NID = JO.SelectToken("event.nodeId").Value<int>();
                 JObject IJO = JO.SelectToken("event.args").Value<JObject>();
@@ -67,7 +102,7 @@ namespace ZWaveJS.NET
                 N.Trigger_ValueNotification(IJO);
             });
 
-            EventMap.Add("notification", (JO) =>
+            NodeEventMap.Add("notification", (JO) =>
             {
                 int NID = JO.SelectToken("event.nodeId").Value<int>();
                 int CCID = JO.SelectToken("event.ccId").Value<int>();
@@ -76,21 +111,28 @@ namespace ZWaveJS.NET
                 N.Trigger_Notification(CCID, IJO);
             });
 
-            EventMap.Add("wake up", (JO) =>
+            NodeEventMap.Add("dead", (JO) =>
+            {
+                int NID = JO.SelectToken("event.nodeId").Value<int>();
+                ZWaveNode N = this.Controller.Nodes.Get(NID);
+                N.Trigger_NodeDead();
+            });
+
+            NodeEventMap.Add("wake up", (JO) =>
             {
                 int NID = JO.SelectToken("event.nodeId").Value<int>();
                 ZWaveNode N = this.Controller.Nodes.Get(NID);
                 N.Trigger_NodeAwake();
             });
 
-            EventMap.Add("sleep", (JO) =>
+            NodeEventMap.Add("sleep", (JO) =>
             {
                 int NID = JO.SelectToken("event.nodeId").Value<int>();
                 ZWaveNode N = this.Controller.Nodes.Get(NID);
                 N.Trigger_NodeAsleep();
             });
 
-            EventMap.Add("ready", (JO) =>
+            NodeEventMap.Add("ready", (JO) =>
             {
                 int NID = JO.SelectToken("event.nodeId").Value<int>();
                 ZWaveNode NNI = JsonConvert.DeserializeObject<ZWaveNode>(JO.SelectToken("event.nodeState").ToString());
@@ -100,21 +142,21 @@ namespace ZWaveJS.NET
                 N.Trigger_NodeReady();
             });
 
-            EventMap.Add("interview started", (JO) =>
+            NodeEventMap.Add("interview started", (JO) =>
             {
                 int NID = JO.SelectToken("event.nodeId").Value<int>();
                 ZWaveNode N = this.Controller.Nodes.Get(NID);
                 N.Trigger_NodeInterviewStarted();
             });
 
-            EventMap.Add("interview completed", (JO) =>
+            NodeEventMap.Add("interview completed", (JO) =>
             {
                 int NID = JO.SelectToken("event.nodeId").Value<int>();
                 ZWaveNode N = this.Controller.Nodes.Get(NID);
                 N.Trigger_NodeInterviewCompleted();
             });
 
-            EventMap.Add("interview failed", (JO) =>
+            NodeEventMap.Add("interview failed", (JO) =>
             {
                 int NID = JO.SelectToken("event.nodeId").Value<int>();
                 JObject IJO = JO.SelectToken("event.args").Value<JObject>();
@@ -125,51 +167,57 @@ namespace ZWaveJS.NET
 
         private void MapControllerEvents()
         {
-            EventMap.Add("inclusion aborted", (JO) =>
+            ControllerEventMap.Add("statistics updated", (JO) =>
+            {
+                ControllerStatistics CS = JsonConvert.DeserializeObject<ControllerStatistics>(JO.SelectToken("event.statistics").ToString());
+                this.Controller.Trigger_StatisticsUpdated(CS);
+            });
+
+            ControllerEventMap.Add("inclusion aborted", (JO) =>
             {
                 this.Controller.Trigger_InclusionAborted();
             });
 
-            EventMap.Add("inclusion started", (JO) =>
+            ControllerEventMap.Add("inclusion started", (JO) =>
             {
                 bool Secure = JO.SelectToken("event.secure").Value<bool>();
                 this.Controller.Trigger_InclusionStarted(Secure);
             });
 
-            EventMap.Add("inclusion stopped", (JO) =>
+            ControllerEventMap.Add("inclusion stopped", (JO) =>
             {
                 this.Controller.Trigger_InclusionStopped();
             });
 
-            EventMap.Add("exclusion started", (JO) =>
+            ControllerEventMap.Add("exclusion started", (JO) =>
             {
                 this.Controller.Trigger_ExclusionStarted();
             });
 
-            EventMap.Add("exclusion stopped", (JO) =>
+            ControllerEventMap.Add("exclusion stopped", (JO) =>
             {
                 this.Controller.Trigger_ExclusionStopped();
             });
 
-            EventMap.Add("node removed", (JO) =>
+            ControllerEventMap.Add("node removed", (JO) =>
             {
                 int NID = JO.SelectToken("event.node.nodeId").Value<int>();
                 this.Controller.Nodes.RemoveNodeFromCollection(NID);
                 this.Controller.Trigger_NodeRemoved(NID);
             });
 
-            EventMap.Add("node added", (JO) =>
+            ControllerEventMap.Add("node added", (JO) =>
             {
                 int NID = JO.SelectToken("event.node.nodeId").Value<int>();
 
                 ZWaveNode NN = new ZWaveNode();
-                NN.nodeId = NID;
+                NN.id = NID;
 
                 this.Controller.Nodes.AddNodeToCollection(NN);
                 this.Controller.Trigger_NodeAdded(NN);
             });
 
-            EventMap.Add("grant security classes", (JO) =>
+            ControllerEventMap.Add("grant security classes", (JO) =>
             {
                 Enums.SecurityClass[] RequestedClasses = JsonConvert.DeserializeObject<Enums.SecurityClass[]>(JO.SelectToken("event.requested.securityClasses").ToString());
                 bool CSA = JO.SelectToken("event.requested.clientSideAuth").Value<bool>();
@@ -185,7 +233,7 @@ namespace ZWaveJS.NET
                 Client.Send(RequestPL);
             });
 
-            EventMap.Add("validate dsk and enter pin", (JO) =>
+            ControllerEventMap.Add("validate dsk and enter pin", (JO) =>
             {
                 string DSK = this.Controller.Trigger_ValidateDSK(JO.SelectToken("event.dsk").Value<string>());
 
@@ -198,13 +246,13 @@ namespace ZWaveJS.NET
                 Client.Send(RequestPL);
             });
 
-            EventMap.Add("heal network progress", (JO) =>
+            ControllerEventMap.Add("heal network progress", (JO) =>
             {
                 Dictionary<string, string> Progress = Newtonsoft.Json.JsonConvert.DeserializeObject<Dictionary<string, string>>(JO.SelectToken("event.progress").ToString());
                 this.Controller.Trigger_HealNetworkProgress(Progress);
             });
 
-            EventMap.Add("heal network done", (JO) =>
+            ControllerEventMap.Add("heal network done", (JO) =>
             {
                 Dictionary<string, string> Result = Newtonsoft.Json.JsonConvert.DeserializeObject<Dictionary<string, string>>(JO.SelectToken("event.result").ToString());
                 this.Controller.Trigger_HealNetworkDone(Result);
@@ -214,8 +262,10 @@ namespace ZWaveJS.NET
 
         private void MapEvents()
         {
-            EventMap = new Dictionary<string, Action<JObject>>();
+            NodeEventMap = new Dictionary<string, Action<JObject>>();
             MapNodeEvents();
+
+            ControllerEventMap = new Dictionary<string, Action<JObject>>();
             MapControllerEvents();
         }
 
@@ -228,8 +278,8 @@ namespace ZWaveJS.NET
             }
 
             Callbacks = new Dictionary<Guid, Action<JObject>>();
-            MapEvents();
             BoolConverter = new CustomBooleanJsonConverter();
+            MapEvents();
 
             Client = new WebsocketClient(Server);
             Client.ReconnectTimeout = null;
@@ -320,12 +370,26 @@ namespace ZWaveJS.NET
 
                 if (Type == "event")
                 {
-                    string EE = JO.SelectToken("event.event").Value<string>();
-                    if (EventMap.ContainsKey(EE))
-                    {
-                        EventMap[EE].Invoke(JO);
-                    }
+                    string _Source = JO.SelectToken("event.source").Value<string>();
+                    string _Event = JO.SelectToken("event.event").Value<string>();
 
+                    switch (_Source)
+                    {
+                        case "node":
+                            if (NodeEventMap.ContainsKey(_Event))
+                            {
+                                NodeEventMap[_Event].Invoke(JO);
+                            }
+                            break;
+
+                        case "controller":
+                            if (ControllerEventMap.ContainsKey(_Event))
+                            {
+                                ControllerEventMap[_Event].Invoke(JO);
+                            }
+                            break;
+
+                    }
                     return;
                 }
             }