+++ Visual Studio Project/ZWaveJS.NET/ZWaveNode.cs
@@ -4,6 +4,7 @@ using System.Threading.Tasks;
 using System.Collections.Generic;
 using Newtonsoft.Json;
 using System.Linq;
+using System.IO;
 namespace ZWaveJS.NET
 {
     public class ZWaveNode
@@ -11,6 +12,28 @@ namespace ZWaveJS.NET
         internal ZWaveNode()
         {
 
+        }
+
+        public delegate void StatisticsUpdatedEvent(ZWaveNode Node, NodeStatistics Statistics);
+        public event StatisticsUpdatedEvent StatisticsUpdated;
+        internal void Trigger_StatisticsUpdated(NodeStatistics Statistics)
+        {
+            this.statistics = Statistics;
+            StatisticsUpdated?.Invoke(this, Statistics);
+        }
+
+        public delegate void FirmwareUpdateFinishedEvent(ZWaveNode Node, int Status, int WaitTime);
+        public event FirmwareUpdateFinishedEvent FirmwareUpdateFinished;
+        internal void Trigger_FirmwareUpdateFinished(int Status, int Time)
+        {
+            FirmwareUpdateFinished?.Invoke(this, Status, Time);
+        }
+
+        public delegate void FirmwareUpdateProgressEvent(ZWaveNode Node, int SentFragments, int TotalFragments);
+        public event FirmwareUpdateProgressEvent FirmwareUpdateProgress;
+        internal void Trigger_FirmwareUpdateProgress(int SentFragments, int TotalFragments)
+        {
+            FirmwareUpdateProgress?.Invoke(this, SentFragments, TotalFragments);
         }
 
         public delegate void ValueNotificationEvent(ZWaveNode Node, JObject Args);
@@ -34,10 +57,19 @@ namespace ZWaveJS.NET
             Notification?.Invoke(this, CCID, Args);
         }
 
+        public delegate void NodeDeadEvent(ZWaveNode Node);
+        public event NodeDeadEvent NodeDead;
+        internal void Trigger_NodeDead()
+        {
+            this.status = Enums.NodeStatus.Dead;
+            NodeDead?.Invoke(this);
+        }
+
         public delegate void AwakeEvent(ZWaveNode Node);
         public event AwakeEvent NodeAwake;
         internal void Trigger_NodeAwake()
         {
+            this.status = Enums.NodeStatus.Awake;
             NodeAwake?.Invoke(this);
         }
 
@@ -45,6 +77,7 @@ namespace ZWaveJS.NET
         public event SleepEvent NodeAsleep;
         internal void Trigger_NodeAsleep()
         {
+            this.status = Enums.NodeStatus.Asleep;
             NodeAsleep?.Invoke(this);
         }
 
@@ -52,6 +85,7 @@ namespace ZWaveJS.NET
         public event NodeReadyEvent NodeReady;
         internal void Trigger_NodeReady()
         {
+            this.ready = true;
             NodeReady?.Invoke(this);
         }
 
@@ -76,6 +110,55 @@ namespace ZWaveJS.NET
             NodeInterviewFailed?.Invoke(this, Args);
         }
 
+
+        public Task<bool> AbortFirmwareUpdate()
+        {
+            Guid ID = Guid.NewGuid();
+
+            TaskCompletionSource<bool> Result = new TaskCompletionSource<bool>();
+            Driver.Callbacks.Add(ID, (JO) =>
+            {
+                Result.SetResult(true);
+            });
+
+            Dictionary<string, object> Request = new Dictionary<string, object>();
+            Request.Add("messageId", ID);
+            Request.Add("command", Enums.Commands.AbortFirmwareUpdate);
+            Request.Add("nodeId", this.id);
+          
+
+            string RequestPL = Newtonsoft.Json.JsonConvert.SerializeObject(Request);
+            Driver.Client.Send(RequestPL);
+
+            return Result.Task;
+        }
+
+        public Task<bool> BeginFirmwareUpdate(string FileName)
+        {
+            Guid ID = Guid.NewGuid();
+
+            TaskCompletionSource<bool> Result = new TaskCompletionSource<bool>();
+            Driver.Callbacks.Add(ID, (JO) =>
+            {
+                Result.SetResult(true);
+            });
+
+            FileInfo FI = new FileInfo(FileName);
+            byte[] FileData = File.ReadAllBytes(FileName);
+
+            Dictionary<string, object> Request = new Dictionary<string, object>();
+            Request.Add("messageId", ID);
+            Request.Add("command", Enums.Commands.BeginFirmwareUpdate);
+            Request.Add("nodeId", this.id);
+            Request.Add("firmwareFile", FileData);
+            Request.Add("firmwareFilename", FI.Name);
+
+            string RequestPL = Newtonsoft.Json.JsonConvert.SerializeObject(Request);
+            Driver.Client.Send(RequestPL);
+
+            return Result.Task;
+        }
+
         public Task<bool> RefreshInfo()
         {
             Guid ID = Guid.NewGuid();
@@ -89,7 +172,7 @@ namespace ZWaveJS.NET
             Dictionary<string, object> Request = new Dictionary<string, object>();
             Request.Add("messageId", ID);
             Request.Add("command", Enums.Commands.RefreshInfo);
-            Request.Add("nodeId", this.nodeId);
+            Request.Add("nodeId", this.id);
 
             string RequestPL = Newtonsoft.Json.JsonConvert.SerializeObject(Request);
             Driver.Client.Send(RequestPL);
@@ -111,7 +194,7 @@ namespace ZWaveJS.NET
             Request.Add("messageId", ID);
             Request.Add("command", Enums.Commands.GetValue);
             Request.Add("valueId", ValueID);
-            Request.Add("nodeId", this.nodeId);
+            Request.Add("nodeId", this.id);
 
             string RequestPL = Newtonsoft.Json.JsonConvert.SerializeObject(Request);
             Driver.Client.Send(RequestPL);
@@ -132,7 +215,7 @@ namespace ZWaveJS.NET
             Dictionary<string, object> Request = new Dictionary<string, object>();
             Request.Add("messageId", ID);
             Request.Add("command", Enums.Commands.SetValue);
-            Request.Add("nodeId", this.nodeId);
+            Request.Add("nodeId", this.id);
             Request.Add("valueId", ValueID);
             Request.Add("value", Value);
 
@@ -160,7 +243,7 @@ namespace ZWaveJS.NET
             Dictionary<string, object> Request = new Dictionary<string, object>();
             Request.Add("messageId", ID);
             Request.Add("command", Enums.Commands.PollValue);
-            Request.Add("nodeId", this.nodeId);
+            Request.Add("nodeId", this.id);
             Request.Add("valueId", ValueID);
 
             string RequestPL = Newtonsoft.Json.JsonConvert.SerializeObject(Request);
@@ -182,7 +265,7 @@ namespace ZWaveJS.NET
             Dictionary<string, object> Request = new Dictionary<string, object>();
             Request.Add("messageId", ID);
             Request.Add("command", Enums.Commands.GetDefinedValueIDs);
-            Request.Add("nodeId", this.nodeId);
+            Request.Add("nodeId", this.id);
 
 
             string RequestPL = JsonConvert.SerializeObject(Request);
@@ -204,9 +287,31 @@ namespace ZWaveJS.NET
             Dictionary<string, object> Request = new Dictionary<string, object>();
             Request.Add("messageId", ID);
             Request.Add("command", Enums.Commands.GetValueMetadata);
-            Request.Add("nodeId", this.nodeId);
+            Request.Add("nodeId", this.id);
             Request.Add("valueId", VID);
 
+
+            string RequestPL = JsonConvert.SerializeObject(Request);
+            Driver.Client.Send(RequestPL);
+
+            return Result.Task;
+        }
+
+        public Task<bool> SupportsCCAPI(int CommandClass)
+        {
+            Guid ID = Guid.NewGuid();
+
+            TaskCompletionSource<bool> Result = new TaskCompletionSource<bool>();
+            Driver.Callbacks.Add(ID, (JO) =>
+            {
+                Result.SetResult(JO.SelectToken("result.supported").Value<bool>());
+            });
+
+            Dictionary<string, object> Request = new Dictionary<string, object>();
+            Request.Add("messageId", ID);
+            Request.Add("command", Enums.Commands.SupportsCCAPI);
+            Request.Add("nodeId", this.id);
+            Request.Add("commandClass", CommandClass);
 
             string RequestPL = JsonConvert.SerializeObject(Request);
             Driver.Client.Send(RequestPL);
@@ -227,11 +332,10 @@ namespace ZWaveJS.NET
             Dictionary<string, object> Request = new Dictionary<string, object>();
             Request.Add("messageId", ID);
             Request.Add("command", Enums.Commands.InvokeCCAPI);
-            Request.Add("nodeId", this.nodeId);
+            Request.Add("nodeId", this.id);
             Request.Add("commandClass", CommandClass);
             Request.Add("methodName", Method);
             Request.Add("args", Params);
-
 
             string RequestPL = JsonConvert.SerializeObject(Request);
             Driver.Client.Send(RequestPL);
@@ -245,17 +349,82 @@ namespace ZWaveJS.NET
             return EP;
         }
 
+        public Endpoint[] GetAllEndpoints()
+        {
+            return endpoints;
+        }
+
+        public Task<int> GetEndpointCount()
+        {
+            Guid ID = Guid.NewGuid();
+
+            TaskCompletionSource<int> Result = new TaskCompletionSource<int>();
+
+            Driver.Callbacks.Add(ID, (JO) =>
+            {
+                Result.SetResult(JO.SelectToken("result.count").Value<int>()) ;
+            });
+
+            Dictionary<string, object> Request = new Dictionary<string, object>();
+            Request.Add("messageId", ID);
+            Request.Add("command", Enums.Commands.GetEndpointCount);
+            Request.Add("nodeId", this.id);
+
+            string RequestPL = JsonConvert.SerializeObject(Request);
+            Driver.Client.Send(RequestPL);
+
+            return Result.Task;
+        }
+
+        public Task<Enums.SecurityClass> GetHighestSecurityClass()
+        {
+            Guid ID = Guid.NewGuid();
+
+            TaskCompletionSource<Enums.SecurityClass> Result = new TaskCompletionSource<Enums.SecurityClass>();
+
+            Driver.Callbacks.Add(ID, (JO) =>
+            {
+                int Value = JO.SelectToken("result.highestSecurityClass").Value<int>();
+                Result.SetResult((Enums.SecurityClass)Value);
+            });
+
+            Dictionary<string, object> Request = new Dictionary<string, object>();
+            Request.Add("messageId", ID);
+            Request.Add("command", Enums.Commands.GetHighestSecurityClass);
+            Request.Add("nodeId", this.id);
+
+            string RequestPL = JsonConvert.SerializeObject(Request);
+            Driver.Client.Send(RequestPL);
+
+            return Result.Task;
+        }
+
+        public Task<bool> HasSecurityClass(Enums.SecurityClass Class)
+        {
+            Guid ID = Guid.NewGuid();
+
+            TaskCompletionSource<bool> Result = new TaskCompletionSource<bool>();
+
+            Driver.Callbacks.Add(ID, (JO) =>
+            {
+                Result.SetResult(JO.SelectToken("result.hasSecurityClass").Value<bool>());
+            });
+
+            Dictionary<string, object> Request = new Dictionary<string, object>();
+            Request.Add("messageId", ID);
+            Request.Add("command", Enums.Commands.HasSecurityClass);
+            Request.Add("nodeId", this.id);
+            Request.Add("securityClass", Class);
+
+            string RequestPL = JsonConvert.SerializeObject(Request);
+            Driver.Client.Send(RequestPL);
+
+            return Result.Task;
+        }
+
         [Newtonsoft.Json.JsonProperty]
         internal Endpoint[] endpoints { get; set; }
 
-        [Newtonsoft.Json.JsonProperty]
-        public int nodeId { get; internal set; }
-        [Newtonsoft.Json.JsonProperty]
-        public int index { get; internal set; }
-        [Newtonsoft.Json.JsonProperty]
-        public int installerIcon { get; internal set; }
-        [Newtonsoft.Json.JsonProperty]
-        public int userIcon { get; internal set; }
         [Newtonsoft.Json.JsonProperty]
         public Enums.NodeStatus status { get; internal set; }
         [Newtonsoft.Json.JsonProperty]
@@ -279,10 +448,6 @@ namespace ZWaveJS.NET
         [Newtonsoft.Json.JsonProperty]
         public DeviceConfig deviceConfig { get; internal set; }
         [Newtonsoft.Json.JsonProperty]
-        public string label { get; internal set; }
-        [Newtonsoft.Json.JsonProperty]
-        public int interviewAttempts { get; internal set; }
-        [Newtonsoft.Json.JsonProperty]
         public object isFrequentListening { get; internal set; }
         [Newtonsoft.Json.JsonProperty]
         public long maxDataRate { get; internal set; }
@@ -295,23 +460,28 @@ namespace ZWaveJS.NET
         [Newtonsoft.Json.JsonProperty]
         public bool supportsSecurity { get; internal set; }
         [Newtonsoft.Json.JsonProperty]
-        public int nodeType { get; internal set; }
-        [Newtonsoft.Json.JsonProperty]
         public int zwavePlusNodeType { get; internal set; }
         [Newtonsoft.Json.JsonProperty]
         public int zwavePlusRoleType { get; internal set; }
         [Newtonsoft.Json.JsonProperty]
         public DeviceClass deviceClass { get; internal set; }
         [Newtonsoft.Json.JsonProperty]
-        public CommandClass[] commandClasses { get; internal set; }
-        [Newtonsoft.Json.JsonProperty]
         public string interviewStage { get; internal set; }
         [Newtonsoft.Json.JsonProperty]
         public string deviceDatabaseUrl { get; internal set; }
         [Newtonsoft.Json.JsonProperty]
-        public int highestSecurityClass { get; internal set; }
+        public int interviewAttempts { get; internal set; }
         [Newtonsoft.Json.JsonProperty]
-        public ValueDump[] values { get; internal set; }
+        public string label { get; internal set; }
+        [Newtonsoft.Json.JsonProperty]
+        public int nodeType { get; internal set; }
+        [Newtonsoft.Json.JsonProperty]
+        public CommandClass[] commandClasses { get; internal set; }
+        [Newtonsoft.Json.JsonProperty]
+        public NodeStatistics statistics { get; internal set; }
+
+        [Newtonsoft.Json.JsonProperty(PropertyName = "nodeId")]
+        public int id { get; internal set; }
 
         private bool _KeepAwake;
         [Newtonsoft.Json.JsonProperty]
@@ -331,7 +501,7 @@ namespace ZWaveJS.NET
                     Dictionary<string, object> Request = new Dictionary<string, object>();
                     Request.Add("messageId", ID);
                     Request.Add("command", Enums.Commands.KeepNodeAwake);
-                    Request.Add("nodeId", this.nodeId);
+                    Request.Add("nodeId", this.id);
                     Request.Add("keepAwake", _KeepAwake);
 
                     string RequestPL = JsonConvert.SerializeObject(Request);
@@ -350,15 +520,15 @@ namespace ZWaveJS.NET
             }
             set
             {
+                _Name = value;
                 if (Driver.Inited)
                 {
-                    _Name = value;
                     Guid ID = Guid.NewGuid();
 
                     Dictionary<string, object> Request = new Dictionary<string, object>();
                     Request.Add("messageId", ID);
                     Request.Add("command", Enums.Commands.SetName);
-                    Request.Add("nodeId", this.nodeId);
+                    Request.Add("nodeId", this.id);
                     Request.Add("name", _Name);
 
                     string RequestPL = JsonConvert.SerializeObject(Request);
@@ -378,15 +548,16 @@ namespace ZWaveJS.NET
             }
             set
             {
+                _Location = value;
                 if (Driver.Inited)
                 {
-                    _Location = value;
+                    
                     Guid ID = Guid.NewGuid();
 
                     Dictionary<string, object> Request = new Dictionary<string, object>();
                     Request.Add("messageId", ID);
                     Request.Add("command", Enums.Commands.SetLocation);
-                    Request.Add("nodeId", this.nodeId);
+                    Request.Add("nodeId", this.id);
                     Request.Add("location", _Location);
 
                     string RequestPL = JsonConvert.SerializeObject(Request);